feat: add setting screen and re route all the screen
